feat: start to rewrite totest utility
